feat(container): revamp cookie deletion
ref: #MANAGER-21211
