+++ packages/manager/apps/container/src/App.tsx
@@ -21,6 +21,7 @@ import { setupDevApplication } from '@/core/dev';
 import { ApplicationProvider } from '@/context';
 import Container from '@/container';
 import { ApiError } from './types/error.type';
+import { deleteMatchingCookies } from './helpers';
 
 function reloadPage() {
   window.location.reload();
@@ -53,26 +54,6 @@ function setupI18n(locale: string) {
       },
       postProcess: 'normalize',
     });
-}
-
-const deleteMatchingCookies = () => {
-  const cookies = document.cookie.split(';');
-  cookies.forEach((cookie) => {
-    const cookieName = cookie.split('=')[0].trim();
-    if (cookieName.startsWith('_biz_') || cookieName.startsWith('_mkto_trk') || cookieName.startsWith('_gcl_au')) {
-      // Delete the cookie by setting its expiration date to the past
-      document.cookie = `${cookieName}=; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT`;
-      // Also try to delete with domain variations
-      const domain = window.location.hostname;
-      document.cookie = `${cookieName}=; path=/; domain=${domain}; expires=Thu, 01 Jan 1970 00:00:00 GMT`;
-      // Try with parent domain
-      const domainParts = domain.split('.');
-      if (domainParts.length > 2) {
-        const parentDomain = domainParts.slice(-2).join('.');
-        document.cookie = `${cookieName}=; path=/; domain=.${parentDomain}; expires=Thu, 01 Jan 1970 00:00:00 GMT`;
-      }
-    }
-  });
 }
 
 const App = () => {
